Moving loading state management to custom hook
